\name{computeTotal}
\alias{computeTotal}
\title{
Compute sum of all units in symData
}
\description{
Compute sum of all units in \code{symData}. Used for detecting significant deviations of variables in specific cluster.
}
\usage{
computeTotal(dataset)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{
       object \code{symData}
}
}
\value{
\code{symObject} - all units summed
}
\author{
  Natasa Kejzar
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }
