\name{popul06m}
\alias{popul06m}
\docType{data}
\title{
Population data for male for 224 World Countries in 2006
}
\description{
Population data for male for 224 world countries in the year 2006.
}
\usage{data("popul06m")}
\format{
  A data frame with 224 observations on the following 17 variables.
  \describe{
    \item{\samp{0-4}}{a numeric vector}
    \item{\samp{5-9}}{a numeric vector}
    \item{\samp{10-14}}{a numeric vector}
    \item{\samp{15-19}}{a numeric vector}
    \item{\samp{20-24}}{a numeric vector}
    \item{\samp{25-29}}{a numeric vector}
    \item{\samp{30-34}}{a numeric vector}
    \item{\samp{35-39}}{a numeric vector}
    \item{\samp{40-44}}{a numeric vector}
    \item{\samp{45-49}}{a numeric vector}
    \item{\samp{50-54}}{a numeric vector}
    \item{\samp{55-59}}{a numeric vector}
    \item{\samp{60-64}}{a numeric vector}
    \item{\samp{65-69}}{a numeric vector}
    \item{\samp{70-74}}{a numeric vector}
    \item{\samp{75-79}}{a numeric vector}
    \item{\samp{80+}}{a numeric vector}
  }
}
\source{
International Data Base. \url{http://www.census.gov/ipc/www/idbnew.html}
Cook Islands Statistics Office. \url{http://www.stats.gov.ck/}
Statistics and Demography. \url{http://www.spc.int/sdp/}
}
\examples{
data(popul06m)
}
\keyword{datasets}
